Moved shared code to absract action container class
